Every function in its own file, part 9 [#CLICKHOUSE-2]
