+++ dbms/src/Functions/acos.cpp
@@ -1,0 +1,15 @@
+#include <Functions/FunctionMathUnaryFloat64.h>
+#include <Functions/FunctionFactory.h>
+
+namespace DB
+{
+
+struct AcosName { static constexpr auto name = "acos"; };
+using FunctionAcos = FunctionMathUnaryFloat64<UnaryFunctionVectorized<AcosName, acos>>;
+
+void registerFunctionAcos(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionAcos>(FunctionFactory::CaseInsensitive);
+}
+
+}
